src/main/java/com/example/eumserver/domain/team/announcement/service/AnnouncementService.java
package com.example.eumserver.domain.team.announcement.service;

import com.example.eumserver.domain.team.announcement.dto.AnnouncementFilter;
import com.example.eumserver.domain.team.announcement.dto.AnnouncementResponse;
import com.example.eumserver.domain.team.announcement.repository.AnnouncementRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageRequest;
import org.springframework.data.domain.Pageable;
import org.springframework.data.domain.Sort;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.ArrayList;
import java.util.List;

@Service
@Transactional(readOnly = true)
@RequiredArgsConstructor
public class AnnouncementService {

    private final AnnouncementRepository announcementRepository;

    public Page<AnnouncementResponse> getFilteredAnnouncementsWithPaging(
            AnnouncementFilter filter,
            int page
    ) {
        List<Sort.Order> sorts = new ArrayList<>();
        sorts.add(Sort.Order.desc("create_data"));
        Pageable pageable = PageRequest.of(page, 12, Sort.by(sorts));
        return announcementRepository.getFilteredAnnouncementsWithPaging(filter, pageable);
    }

}
